Refactor endpoints into routers
